backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
lightgbm
scipy
slowapi
python-dotenv
python-multipart
nltk
rich
typer
starlette